// frontend/src/backend.d.ts
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
// ...
export interface ProductCreate {
    id: string;
    media: ProductMedia;
    inStock: boolean;
    name: string;
    description: string;
    gender: Gender;
    category: string;
    priceInCents: bigint;
}
export interface ProductMedia {
    video?: ExternalBlob;
    images: Array<ExternalBlob>;
}
export interface CancelReason {
    reason: string;
}
export interface Product {
    id: string;
    media: ProductMedia;
    inStock: boolean;
    name: string;
    description: string;
    gender: Gender;
    category: string;
    priceInCents: bigint;
}
export enum Gender {
    boys = "boys",
    girls = "girls"
}